도메인 미들네임 변경 example -> JDS
